changed tickintervall to float for faster games

+++ Assets/Scripts/Utility/TickManager.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class TickManager : MonoBehaviour
 {
     [SerializeField] GlobalInformation global;
-    [SerializeField] int intervall = 1;
+    [SerializeField] float intervall = 1;
 
     bool pause;
